Step 5c (drift fix): move spec/model/ → engine/spec/model/; delete top-level spec/
Per DIR-PACKAGE-DRIFT-FIX-punit step 5c. The internal spec model
classes (ExecutionSpecification + nested types, LatencyBaseline,
SuccessCriteria, SpecificationValidationException) move under
engine/spec/model/. After this commit the top-level spec/ package
is empty and is deleted. The author-facing api/spec/ surface stays
public and untouched.

Touched:
- git mv 4 main files from spec/model/ to engine/spec/model/.
- Rewrite package declarations.
- Bulk-rewrite import paths across all four punit modules and
  punitexamples.
- Delete top-level spec/package-info.java and the now-empty spec/
  directories (main and test).
- Drop the spec.* row from docs/DEVELOPER-API.md's
  package-visibility table.

Defect fix in the same commit:
- The bulk import-rewrite sweeps in steps 5a/5b/5c had inadvertently
  rewritten the literal package strings inside
  PackageStructureArchitectureTest.noTopLevelSpecRuntimePackages's
  rule definition — turning the rule into one that forbids the new
  (target) location instead of the old. Restored the rule's
  package-list strings to the original old-package names. The rule
  now correctly verifies that the old top-level spec/ runtime
  packages have no current residents.
- The corrupted intermediate rule-IDs (60dd59f5..., e1339561...,
  ef223fdf...) created by the freeze framework against the broken
  rule are deleted from archunit_store/ and stored.rules.

Architecture-test bookkeeping:
- noTopLevelSpecRuntimePackages rule's freeze-store: empty — at
  target state.
- engineUnderInternal rule's freeze-store: refrozen to absorb
  engine.spec.model residents (intermediate state).
- Other rules' stores: unchanged.

Tests green across all four punit modules; punitexamples composite
build green.

Closes ArchUnit rule: noTopLevelSpecRuntimePackages.
