+++ punit-core/archunit_store/b705f7c8-ab73-4b68-8e35-130a0cf7f515
@@ -56,6 +56,18 @@ Class <org.javai.punit.engine.spec.expiration.ExpirationEvaluator> does reside i
 Class <org.javai.punit.engine.spec.expiration.ExpirationReportPublisher> does reside in a package 'org.javai.punit.engine..' in (ExpirationReportPublisher.java:0)
 Class <org.javai.punit.engine.spec.expiration.ExpirationWarningRenderer> does reside in a package 'org.javai.punit.engine..' in (ExpirationWarningRenderer.java:0)
 Class <org.javai.punit.engine.spec.expiration.WarningLevel> does reside in a package 'org.javai.punit.engine..' in (WarningLevel.java:0)
+Class <org.javai.punit.engine.spec.model.ExecutionSpecification$Builder> does reside in a package 'org.javai.punit.engine..' in (ExecutionSpecification.java:0)
+Class <org.javai.punit.engine.spec.model.ExecutionSpecification$CostEnvelope> does reside in a package 'org.javai.punit.engine..' in (ExecutionSpecification.java:0)
+Class <org.javai.punit.engine.spec.model.ExecutionSpecification$EmpiricalBasis> does reside in a package 'org.javai.punit.engine..' in (ExecutionSpecification.java:0)
+Class <org.javai.punit.engine.spec.model.ExecutionSpecification$ExtendedStatistics> does reside in a package 'org.javai.punit.engine..' in (ExecutionSpecification.java:0)
+Class <org.javai.punit.engine.spec.model.ExecutionSpecification$FactorSourceMetadata> does reside in a package 'org.javai.punit.engine..' in (ExecutionSpecification.java:0)
+Class <org.javai.punit.engine.spec.model.ExecutionSpecification$SpecRequirements> does reside in a package 'org.javai.punit.engine..' in (ExecutionSpecification.java:0)
+Class <org.javai.punit.engine.spec.model.ExecutionSpecification> does reside in a package 'org.javai.punit.engine..' in (ExecutionSpecification.java:0)
+Class <org.javai.punit.engine.spec.model.ExpressionSuccessCriteria> does reside in a package 'org.javai.punit.engine..' in (SuccessCriteria.java:0)
+Class <org.javai.punit.engine.spec.model.LatencyBaseline> does reside in a package 'org.javai.punit.engine..' in (LatencyBaseline.java:0)
+Class <org.javai.punit.engine.spec.model.SpecificationValidationException> does reside in a package 'org.javai.punit.engine..' in (SpecificationValidationException.java:0)
+Class <org.javai.punit.engine.spec.model.SuccessCriteria$1> does reside in a package 'org.javai.punit.engine..' in (SuccessCriteria.java:0)
+Class <org.javai.punit.engine.spec.model.SuccessCriteria> does reside in a package 'org.javai.punit.engine..' in (SuccessCriteria.java:0)
 Class <org.javai.punit.engine.spec.registry.LayeredSpecRepository> does reside in a package 'org.javai.punit.engine..' in (LayeredSpecRepository.java:0)
 Class <org.javai.punit.engine.spec.registry.SpecRepository> does reside in a package 'org.javai.punit.engine..' in (SpecRepository.java:0)
 Class <org.javai.punit.engine.spec.registry.SpecSchemaValidator$ValidationResult> does reside in a package 'org.javai.punit.engine..' in (SpecSchemaValidator.java:0)
