--- punit-core/archunit_store/e1339561-1860-4b31-875b-cd5455336e71
@@ -1,25 +1,0 @@
-Class <org.javai.punit.engine.spec.registry.LayeredSpecRepository> does reside in any package ['org.javai.punit.spec', 'org.javai.punit.engine.spec.registry..', 'org.javai.punit.spec.expiration..', 'org.javai.punit.spec.model..', 'org.javai.punit.spec.criteria..'] in (LayeredSpecRepository.java:0)
-Class <org.javai.punit.engine.spec.registry.SpecRepository> does reside in any package ['org.javai.punit.spec', 'org.javai.punit.engine.spec.registry..', 'org.javai.punit.spec.expiration..', 'org.javai.punit.spec.model..', 'org.javai.punit.spec.criteria..'] in (SpecRepository.java:0)
-Class <org.javai.punit.engine.spec.registry.SpecSchemaValidator$ValidationResult> does reside in any package ['org.javai.punit.spec', 'org.javai.punit.engine.spec.registry..', 'org.javai.punit.spec.expiration..', 'org.javai.punit.spec.model..', 'org.javai.punit.spec.criteria..'] in (SpecSchemaValidator.java:0)
-Class <org.javai.punit.engine.spec.registry.SpecSchemaValidator> does reside in any package ['org.javai.punit.spec', 'org.javai.punit.engine.spec.registry..', 'org.javai.punit.spec.expiration..', 'org.javai.punit.spec.model..', 'org.javai.punit.spec.criteria..'] in (SpecSchemaValidator.java:0)
-Class <org.javai.punit.engine.spec.registry.SpecificationIntegrityException> does reside in any package ['org.javai.punit.spec', 'org.javai.punit.engine.spec.registry..', 'org.javai.punit.spec.expiration..', 'org.javai.punit.spec.model..', 'org.javai.punit.spec.criteria..'] in (SpecificationIntegrityException.java:0)
-Class <org.javai.punit.engine.spec.registry.SpecificationLoadException> does reside in any package ['org.javai.punit.spec', 'org.javai.punit.engine.spec.registry..', 'org.javai.punit.spec.expiration..', 'org.javai.punit.spec.model..', 'org.javai.punit.spec.criteria..'] in (SpecificationLoadException.java:0)
-Class <org.javai.punit.engine.spec.registry.SpecificationLoader> does reside in any package ['org.javai.punit.spec', 'org.javai.punit.engine.spec.registry..', 'org.javai.punit.spec.expiration..', 'org.javai.punit.spec.model..', 'org.javai.punit.spec.criteria..'] in (SpecificationLoader.java:0)
-Class <org.javai.punit.engine.spec.registry.SpecificationNotFoundException> does reside in any package ['org.javai.punit.spec', 'org.javai.punit.engine.spec.registry..', 'org.javai.punit.spec.expiration..', 'org.javai.punit.spec.model..', 'org.javai.punit.spec.criteria..'] in (SpecificationNotFoundException.java:0)
-Class <org.javai.punit.engine.spec.registry.SpecificationRegistry> does reside in any package ['org.javai.punit.spec', 'org.javai.punit.engine.spec.registry..', 'org.javai.punit.spec.expiration..', 'org.javai.punit.spec.model..', 'org.javai.punit.spec.criteria..'] in (SpecificationRegistry.java:0)
-Class <org.javai.punit.spec.expiration.ExpirationEvaluator> does reside in any package ['org.javai.punit.spec', 'org.javai.punit.engine.spec.registry..', 'org.javai.punit.spec.expiration..', 'org.javai.punit.spec.model..', 'org.javai.punit.spec.criteria..'] in (ExpirationEvaluator.java:0)
-Class <org.javai.punit.spec.expiration.ExpirationReportPublisher> does reside in any package ['org.javai.punit.spec', 'org.javai.punit.engine.spec.registry..', 'org.javai.punit.spec.expiration..', 'org.javai.punit.spec.model..', 'org.javai.punit.spec.criteria..'] in (ExpirationReportPublisher.java:0)
-Class <org.javai.punit.spec.expiration.ExpirationWarningRenderer> does reside in any package ['org.javai.punit.spec', 'org.javai.punit.engine.spec.registry..', 'org.javai.punit.spec.expiration..', 'org.javai.punit.spec.model..', 'org.javai.punit.spec.criteria..'] in (ExpirationWarningRenderer.java:0)
-Class <org.javai.punit.spec.expiration.WarningLevel> does reside in any package ['org.javai.punit.spec', 'org.javai.punit.engine.spec.registry..', 'org.javai.punit.spec.expiration..', 'org.javai.punit.spec.model..', 'org.javai.punit.spec.criteria..'] in (WarningLevel.java:0)
-Class <org.javai.punit.spec.model.ExecutionSpecification$Builder> does reside in any package ['org.javai.punit.spec', 'org.javai.punit.engine.spec.registry..', 'org.javai.punit.spec.expiration..', 'org.javai.punit.spec.model..', 'org.javai.punit.spec.criteria..'] in (ExecutionSpecification.java:0)
-Class <org.javai.punit.spec.model.ExecutionSpecification$CostEnvelope> does reside in any package ['org.javai.punit.spec', 'org.javai.punit.engine.spec.registry..', 'org.javai.punit.spec.expiration..', 'org.javai.punit.spec.model..', 'org.javai.punit.spec.criteria..'] in (ExecutionSpecification.java:0)
-Class <org.javai.punit.spec.model.ExecutionSpecification$EmpiricalBasis> does reside in any package ['org.javai.punit.spec', 'org.javai.punit.engine.spec.registry..', 'org.javai.punit.spec.expiration..', 'org.javai.punit.spec.model..', 'org.javai.punit.spec.criteria..'] in (ExecutionSpecification.java:0)
-Class <org.javai.punit.spec.model.ExecutionSpecification$ExtendedStatistics> does reside in any package ['org.javai.punit.spec', 'org.javai.punit.engine.spec.registry..', 'org.javai.punit.spec.expiration..', 'org.javai.punit.spec.model..', 'org.javai.punit.spec.criteria..'] in (ExecutionSpecification.java:0)
-Class <org.javai.punit.spec.model.ExecutionSpecification$FactorSourceMetadata> does reside in any package ['org.javai.punit.spec', 'org.javai.punit.engine.spec.registry..', 'org.javai.punit.spec.expiration..', 'org.javai.punit.spec.model..', 'org.javai.punit.spec.criteria..'] in (ExecutionSpecification.java:0)
-Class <org.javai.punit.spec.model.ExecutionSpecification$SpecRequirements> does reside in any package ['org.javai.punit.spec', 'org.javai.punit.engine.spec.registry..', 'org.javai.punit.spec.expiration..', 'org.javai.punit.spec.model..', 'org.javai.punit.spec.criteria..'] in (ExecutionSpecification.java:0)
-Class <org.javai.punit.spec.model.ExecutionSpecification> does reside in any package ['org.javai.punit.spec', 'org.javai.punit.engine.spec.registry..', 'org.javai.punit.spec.expiration..', 'org.javai.punit.spec.model..', 'org.javai.punit.spec.criteria..'] in (ExecutionSpecification.java:0)
-Class <org.javai.punit.spec.model.ExpressionSuccessCriteria> does reside in any package ['org.javai.punit.spec', 'org.javai.punit.engine.spec.registry..', 'org.javai.punit.spec.expiration..', 'org.javai.punit.spec.model..', 'org.javai.punit.spec.criteria..'] in (SuccessCriteria.java:0)
-Class <org.javai.punit.spec.model.LatencyBaseline> does reside in any package ['org.javai.punit.spec', 'org.javai.punit.engine.spec.registry..', 'org.javai.punit.spec.expiration..', 'org.javai.punit.spec.model..', 'org.javai.punit.spec.criteria..'] in (LatencyBaseline.java:0)
-Class <org.javai.punit.spec.model.SpecificationValidationException> does reside in any package ['org.javai.punit.spec', 'org.javai.punit.engine.spec.registry..', 'org.javai.punit.spec.expiration..', 'org.javai.punit.spec.model..', 'org.javai.punit.spec.criteria..'] in (SpecificationValidationException.java:0)
-Class <org.javai.punit.spec.model.SuccessCriteria$1> does reside in any package ['org.javai.punit.spec', 'org.javai.punit.engine.spec.registry..', 'org.javai.punit.spec.expiration..', 'org.javai.punit.spec.model..', 'org.javai.punit.spec.criteria..'] in (SuccessCriteria.java:0)
-Class <org.javai.punit.spec.model.SuccessCriteria> does reside in any package ['org.javai.punit.spec', 'org.javai.punit.engine.spec.registry..', 'org.javai.punit.spec.expiration..', 'org.javai.punit.spec.model..', 'org.javai.punit.spec.criteria..'] in (SuccessCriteria.java:0)
