+++ punit-core/archunit_store/stored.rules
@@ -1,5 +1,5 @@
 #
-#Sun May 10 20:29:37 CEST 2026
+#Sun May 10 20:31:54 CEST 2026
 classes\ that\ reside\ in\ a\ package\ 'org.javai.punit.runtime'\ and\ are\ top\ level\ classes\ should\ have\ simple\ name\ 'PUnit',\ because\ runtime/\ is\ part\ of\ the\ public\ surface;\ only\ PUnit\ is\ public\ there.\ Emitters\ and\ resolvers\ move\ to\ internal.runtime/\ under\ DIR-INTERNAL-NAMESPACE-punit\ step\ 4=aac540b7-bc27-4122-9e03-6ddafb8ee35a
 no\ classes\ should\ reside\ in\ a\ package\ 'org.javai.punit.contract..',\ because\ the\ parallel\ and\ unused\ contract/\ stack\ is\ deleted\ in\ full\ under\ DIR-CONTRACT-PACKAGE-REMOVAL-punit=da9be0d8-75d6-46e0-977a-6687262b70fe
 no\ classes\ should\ reside\ in\ a\ package\ 'org.javai.punit.controls..',\ because\ controls/\ collapses;\ budget\ and\ pacing\ move\ under\ engine/\ (then\ internal.engine/)\ under\ DIR-PACKAGE-DRIFT-FIX-punit\ step\ 4=cd07f984-c9bc-473f-8624-454ace84b950
@@ -9,8 +9,6 @@ no\ classes\ should\ reside\ in\ a\ package\ 'org.javai.punit.model..',\ because
 no\ classes\ should\ reside\ in\ a\ package\ 'org.javai.punit.power..',\ because\ power/\ is\ a\ single-class\ package;\ its\ contents\ fold\ into\ statistics/\ under\ DIR-PACKAGE-DRIFT-FIX-punit\ step\ 1=caa1fdee-320c-411c-ba6e-f62d660054c9
 no\ classes\ should\ reside\ in\ a\ package\ 'org.javai.punit.reporting..',\ because\ reporting/\ (sink\ implementations\ and\ renderers)\ relocates\ under\ internal.reporting/\ under\ DIR-INTERNAL-NAMESPACE-punit\ step\ 2.\ The\ VerdictSink\ interface\ stays\ in\ verdict/\ \u2014\ only\ implementations\ move.=8340eb1d-f56f-49d7-853a-f8267e571e40
 no\ classes\ should\ reside\ in\ a\ package\ 'org.javai.punit.util..',\ because\ util/\ relocates\ under\ internal.util/\ under\ DIR-INTERNAL-NAMESPACE-punit\ step\ 3=239193ee-09df-4395-aa00-c7cabac03227
-no\ classes\ should\ reside\ in\ any\ package\ ['org.javai.punit.spec',\ 'org.javai.punit.engine.spec.registry..',\ 'org.javai.punit.engine.spec.expiration..',\ 'org.javai.punit.spec.model..',\ 'org.javai.punit.spec.criteria..'],\ because\ runtime\ spec\ materialisation\ relocates\ under\ engine/spec/\ under\ DIR-PACKAGE-DRIFT-FIX-punit\ step\ 5;\ api.spec/\ is\ the\ public\ typed-spec\ surface\ and\ is\ unaffected=60dd59f5-20a3-4e7e-9c0e-98a5a01c666e
-no\ classes\ should\ reside\ in\ any\ package\ ['org.javai.punit.spec',\ 'org.javai.punit.engine.spec.registry..',\ 'org.javai.punit.spec.expiration..',\ 'org.javai.punit.spec.model..',\ 'org.javai.punit.spec.criteria..'],\ because\ runtime\ spec\ materialisation\ relocates\ under\ engine/spec/\ under\ DIR-PACKAGE-DRIFT-FIX-punit\ step\ 5;\ api.spec/\ is\ the\ public\ typed-spec\ surface\ and\ is\ unaffected=e1339561-1860-4b31-875b-cd5455336e71
 no\ classes\ should\ reside\ in\ any\ package\ ['org.javai.punit.spec',\ 'org.javai.punit.spec.registry..',\ 'org.javai.punit.spec.expiration..',\ 'org.javai.punit.spec.model..',\ 'org.javai.punit.spec.criteria..'],\ because\ runtime\ spec\ materialisation\ relocates\ under\ engine/spec/\ under\ DIR-PACKAGE-DRIFT-FIX-punit\ step\ 5;\ api.spec/\ is\ the\ public\ typed-spec\ surface\ and\ is\ unaffected=357246fe-57ce-41da-93fd-bdfe29e38e41
 no\ classes\ that\ reside\ in\ a\ package\ 'org.javai.punit.api'\ should\ depend\ on\ classes\ that\ reside\ outside\ of\ packages\ ['java..',\ 'javax..',\ 'org.javai.punit.api..',\ 'org.javai.outcome..',\ 'org.junit.jupiter.api..'],\ because\ the\ api\ root\ carries\ the\ abstraction\ layer;\ concrete\ implementations\ and\ their\ third-party\ dependencies\ live\ in\ api\ subpackages\ (api.spec,\ api.covariate,\ api.match,\ \u2026)=78a3bd1f-a2f0-4720-b9a1-397a27caa433
 no\ classes\ that\ reside\ outside\ of\ package\ 'org.javai.punit.internal..'\ should\ depend\ on\ classes\ that\ reside\ in\ a\ package\ 'org.javai.punit.internal..',\ because\ the\ internal/\ namespace\ is\ the\ structural\ marker\ for\ framework\ internals;\ consumers\ must\ not\ reach\ in.\ Enforced\ after\ DIR-INTERNAL-NAMESPACE-punit\ lands.=5597aef5-7da4-454d-9b90-c35aa63c63b7
