+++ punit-core/src/main/java/org/javai/punit/engine/spec/registry/LayeredSpecRepository.java
@@ -8,7 +8,7 @@ import java.util.Collections;
 import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
-import org.javai.punit.spec.model.ExecutionSpecification;
+import org.javai.punit.engine.spec.model.ExecutionSpecification;
 
 /**
  * A {@link SpecRepository} that resolves specifications by trying multiple
