+++ punit-core/src/test/java/org/javai/punit/architecture/PackageStructureArchitectureTest.java
@@ -137,8 +137,8 @@ class PackageStructureArchitectureTest {
             ArchRule rule = noClasses()
                     .should().resideInAnyPackage(
                             "org.javai.punit.spec",
-                            "org.javai.punit.engine.spec.registry..",
-                            "org.javai.punit.engine.spec.expiration..",
+                            "org.javai.punit.spec.registry..",
+                            "org.javai.punit.spec.expiration..",
                             "org.javai.punit.spec.model..",
                             "org.javai.punit.spec.criteria..")
                     .because("runtime spec materialisation relocates under engine/spec/ "
